requirements.txt
django>=3.0.0,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0 --no-binary psycopg2

openpyxl
tablib[xlsx]
django-import-export

flake8>=3.6.0,<3.7.0
